updated dummy report to print only for thread 0 and ftp reconstruction

+++ src/processing.h
@@ -283,7 +283,7 @@ typedef struct ftp_session_attr_struct {
     char * packet_request;
     char * response_value;
     uint32_t file_size;
-    uint32_t data_type;
+    uint8_t data_type;
     uint64_t response_time;
     uint8_t data_response_time_seen;
     char * location;
